+++ AtlyssDiscordRichPresence/Display.cs
@@ -9,6 +9,9 @@ public class Display
     {
         Custom,
         Default,
+        Detailed,
+        DetailedPercentages,
+        Pon,
         Emojis
     }
 
@@ -18,10 +21,13 @@ public class Display
         PlayerAlive,
         PlayerDead,
         MainMenu,
+        CharacterCreation,
+        CharacterCreationDetails,
         Exploring,
         Idle,
         FightingInArena,
         FightingBoss,
+        DungeonBossEnd,
         Singleplayer,
         Multiplayer,
 
@@ -38,60 +44,69 @@ public class Display
         ClassNovice,
         ClassFighter,
         ClassMystic,
-        ClassBandit
+        ClassBandit,
+
+        // Class tier variables for Fighter
+        ClassTierBerserker,
+        ClassTierPaladin,
+
+        // Class tier variables for Mystic
+        ClassTierMagus,
+        ClassTierBishop,
+
+        // Class tier variables for Bandit
+        ClassTierRogue,
+        ClassTierEngineer
     }
 
     public Display(ConfigFile config)
     {
+        void RegisterCustomPresetText(Texts text, string description)
+        {
+            config.Bind("Display", text.ToString(), GetDefaultPresetText(text), description);
+        }
+
         DisplayPreset = config.Bind("General", nameof(DisplayPreset), DisplayPresets.Custom, "Preset to use for texts. \"Custom\" will use the custom texts defined in the config.");
 
-        PlayerAliveCustom = config.Bind("Display", "PlayerAlive", GetDefaultPresetText(Texts.PlayerAlive), "Text to display for player stats while alive.");
-        PlayerDeadCustom = config.Bind("Display", "PlayerDead", GetDefaultPresetText(Texts.PlayerDead), "Text to display for player stats while dead.");
-        MainMenuCustom = config.Bind("Display", "MainMenu", GetDefaultPresetText(Texts.MainMenu), "Text to display while you're in the main menu.");
-        ExploringCustom = config.Bind("Display", "Exploring", GetDefaultPresetText(Texts.Exploring), "Text to display while exploring the world.");
-        IdleCustom = config.Bind("Display", "Idle", GetDefaultPresetText(Texts.Idle), "Text to display while being idle in the world.");
-        FightingInArenaCustom = config.Bind("Display", "FightingInArena", GetDefaultPresetText(Texts.FightingInArena), "Text to display while an arena is active.");
-        FightingBossCustom = config.Bind("Display", "FightingBoss", GetDefaultPresetText(Texts.FightingBoss), "Text to display while a boss is active.");
-        SingleplayerCustom = config.Bind("Display", "Singleplayer", GetDefaultPresetText(Texts.Singleplayer), "Text to display while in singleplayer.");
-        MultiplayerCustom = config.Bind("Display", "Multiplayer", GetDefaultPresetText(Texts.Multiplayer), "Text to display while in multiplayer.");
+        // Regular texts
+        RegisterCustomPresetText(Texts.PlayerAlive, "Text to display for player stats while alive.");
+        RegisterCustomPresetText(Texts.PlayerDead, "Text to display for player stats while dead.");
+        RegisterCustomPresetText(Texts.MainMenu, "Text to display while you're in the main menu.");
+        RegisterCustomPresetText(Texts.CharacterCreation, "Text to display while you're in the character creation screen.");
+        RegisterCustomPresetText(Texts.CharacterCreationDetails, "Text to display about the current character you're creating in the creation screen.");
+        RegisterCustomPresetText(Texts.Exploring, "Text to display while exploring the world.");
+        RegisterCustomPresetText(Texts.Idle, "Text to display while being idle in the world.");
+        RegisterCustomPresetText(Texts.FightingInArena, "Text to display while an arena is active.");
+        RegisterCustomPresetText(Texts.FightingBoss, "Text to display while a boss is active.");
+        RegisterCustomPresetText(Texts.DungeonBossEnd, "Text to display during the post-boss state of a dungeon.");
+        RegisterCustomPresetText(Texts.Singleplayer, "Text to display while in singleplayer.");
+        RegisterCustomPresetText(Texts.Multiplayer, "Text to display while in multiplayer.");
 
-        RaceImpCustom = config.Bind("Display", "RaceImp", GetDefaultPresetText(Texts.RaceImp), $"Text to use when displaying {PLAYERRACE} (for Imps).");
-        RacePoonCustom = config.Bind("Display", "RacePoon", GetDefaultPresetText(Texts.RacePoon), $"Text to use when displaying {PLAYERRACE} (for Poons).");
-        RaceKuboldCustom = config.Bind("Display", "RaceKubold", GetDefaultPresetText(Texts.RaceKubold), $"Text to use when displaying {PLAYERRACE} (for Kubolds).");
-        RaceByrdleCustom = config.Bind("Display", "RaceByrdle", GetDefaultPresetText(Texts.RaceByrdle), $"Text to use when displaying {PLAYERRACE} (for Byrdles).");
-        RaceChangCustom = config.Bind("Display", "RaceChang", GetDefaultPresetText(Texts.RaceChang), $"Text to use when displaying {PLAYERRACE} (for Changs).");
+        // Race names
+        RegisterCustomPresetText(Texts.RaceImp, $"Text to use when displaying {PLAYERRACE} (for Imps).");
+        RegisterCustomPresetText(Texts.RacePoon, $"Text to use when displaying {PLAYERRACE} (for Poons).");
+        RegisterCustomPresetText(Texts.RaceKubold, $"Text to use when displaying {PLAYERRACE} (for Kubolds).");
+        RegisterCustomPresetText(Texts.RaceByrdle, $"Text to use when displaying {PLAYERRACE} (for Byrdles).");
+        RegisterCustomPresetText(Texts.RaceChang, $"Text to use when displaying {PLAYERRACE} (for Changs).");
 
-        ClassNoviceCustom = config.Bind("Display", "ClassNovice", GetDefaultPresetText(Texts.ClassNovice), $"Text to use when displaying {PLAYERCLASS} (for Novices).");
-        ClassFighterCustom = config.Bind("Display", "ClassFighter", GetDefaultPresetText(Texts.ClassFighter), $"Text to use when displaying {PLAYERCLASS} (for Fighters).");
-        ClassMysticCustom = config.Bind("Display", "ClassMystic", GetDefaultPresetText(Texts.ClassMystic), $"Text to use when displaying {PLAYERCLASS} (for Mystics).");
-        ClassBanditCustom = config.Bind("Display", "ClassBandit", GetDefaultPresetText(Texts.ClassBandit), $"Text to use when displaying {PLAYERCLASS} (for Bandits).");
+        // Class names
+        RegisterCustomPresetText(Texts.ClassNovice, $"Text to use when displaying {PLAYERCLASS} (for Novices).");
+        RegisterCustomPresetText(Texts.ClassFighter, $"Text to use when displaying {PLAYERCLASS} (for Fighters).");
+        RegisterCustomPresetText(Texts.ClassMystic, $"Text to use when displaying {PLAYERCLASS} (for Mystics).");
+        RegisterCustomPresetText(Texts.ClassBandit, $"Text to use when displaying {PLAYERCLASS} (for Bandits).");
+
+        // Class tier names
+        RegisterCustomPresetText(Texts.ClassTierBerserker, $"Text to use when displaying {PLAYERCLASS} (for Berserkers - Fighter tier).");
+        RegisterCustomPresetText(Texts.ClassTierPaladin, $"Text to use when displaying {PLAYERCLASS} (for Paladins - Fighter tier).");
+        RegisterCustomPresetText(Texts.ClassTierMagus, $"Text to use when displaying {PLAYERCLASS} (for Magi - Mystic tier).");
+        RegisterCustomPresetText(Texts.ClassTierBishop, $"Text to use when displaying {PLAYERCLASS} (for Bishops - Mystic tier).");
+        RegisterCustomPresetText(Texts.ClassTierRogue, $"Text to use when displaying {PLAYERCLASS} (for Rogues - Bandit tier).");
+        RegisterCustomPresetText(Texts.ClassTierEngineer, $"Text to use when displaying {PLAYERCLASS} (for Engineers - Bandit tier).");
     }
 
     public ConfigEntry<DisplayPresets> DisplayPreset { get; }
 
-    // Full texts (Custom)
-    public ConfigEntry<string> PlayerAliveCustom { get; }
-    public ConfigEntry<string> PlayerDeadCustom { get; }
-    public ConfigEntry<string> MainMenuCustom { get; }
-    public ConfigEntry<string> ExploringCustom { get; }
-    public ConfigEntry<string> IdleCustom { get; }
-    public ConfigEntry<string> FightingInArenaCustom { get; }
-    public ConfigEntry<string> FightingBossCustom { get; }
-    public ConfigEntry<string> SingleplayerCustom { get; }
-    public ConfigEntry<string> MultiplayerCustom { get; }
-
-    // Race variables (Custom)
-    public ConfigEntry<string> RaceImpCustom { get; }
-    public ConfigEntry<string> RacePoonCustom { get; }
-    public ConfigEntry<string> RaceKuboldCustom { get; }
-    public ConfigEntry<string> RaceByrdleCustom { get; }
-    public ConfigEntry<string> RaceChangCustom { get; }
-
-    // Class variables (Custom)
-    public ConfigEntry<string> ClassNoviceCustom { get; }
-    public ConfigEntry<string> ClassFighterCustom { get; }
-    public ConfigEntry<string> ClassMysticCustom { get; }
-    public ConfigEntry<string> ClassBanditCustom { get; }
+    private readonly Dictionary<Texts, ConfigEntry<string>> _customPresetData  = [];
 
     private static string EscapeVars(string input)
     {
@@ -122,71 +137,68 @@ public class Display
         return input;
     }
 
+    public string GetText(Texts text, GameState state) => ReplaceVars(GetRawText(text), state);
 
-    public string GetVariable(string variable, string value) => variable switch
+    private string GetVariable(string variable, string value) => variable switch
     {
         PLAYERCLASS => GetMappedText(value),
         PLAYERRACE => GetMappedText(value),
+        CHARCREATERACE => GetMappedText(value),
         PLAYERRACEANDCLASS => string.Join(" ", value.Split(" ").Select(GetMappedText)),
         _ => value
     };
 
-    public string GetMappedText(string str) => str.ToLower() switch
+    private string GetMappedText(string str) => str.ToLower() switch
     {
-        "imp" => GetText(Texts.RaceImp),
-        "poon" => GetText(Texts.RacePoon),
-        "kubold" => GetText(Texts.RaceKubold),
-        "byrdle" => GetText(Texts.RaceByrdle),
-        "chang" => GetText(Texts.RaceChang),
-        "novice" => GetText(Texts.ClassNovice),
-        "fighter" => GetText(Texts.ClassFighter),
-        "mystic" => GetText(Texts.ClassMystic),
-        "bandit" => GetText(Texts.ClassBandit),
+        "imp" => GetRawText(Texts.RaceImp),
+        "poon" => GetRawText(Texts.RacePoon),
+        "kubold" => GetRawText(Texts.RaceKubold),
+        "byrdle" => GetRawText(Texts.RaceByrdle),
+        "chang" => GetRawText(Texts.RaceChang),
+        "novice" => GetRawText(Texts.ClassNovice),
+        "fighter" => GetRawText(Texts.ClassFighter),
+        "mystic" => GetRawText(Texts.ClassMystic),
+        "bandit" => GetRawText(Texts.ClassBandit),
+        "berserker" => GetRawText(Texts.ClassTierBerserker),
+        "paladin" => GetRawText(Texts.ClassTierPaladin),
+        "magus" => GetRawText(Texts.ClassTierMagus),
+        "bishop" => GetRawText(Texts.ClassTierBishop),
+        "rogue" => GetRawText(Texts.ClassTierRogue),
+        "engineer" => GetRawText(Texts.ClassTierEngineer),
         _ => str
     };
 
-    public string GetText(Texts text) => DisplayPreset.Value switch
+    private string GetRawText(Texts text) => DisplayPreset.Value switch
     {
         DisplayPresets.Custom => GetCustomPresetText(text),
         DisplayPresets.Default => GetDefaultPresetText(text),
+        DisplayPresets.Detailed => GetDetailedPresetText(text),
+        DisplayPresets.DetailedPercentages => GetDetailedPercentagesPresetText(text),
+        DisplayPresets.Pon => GetPonPresetText(text),
         DisplayPresets.Emojis => GetEmojisPresetText(text),
         _ => "[Unknown Preset]"
     };
 
-    private string GetCustomPresetText(Texts text) => text switch
+    private string GetCustomPresetText(Texts text)
     {
-        Texts.PlayerAlive => PlayerAliveCustom.Value,
-        Texts.PlayerDead => PlayerDeadCustom.Value,
-        Texts.MainMenu => MainMenuCustom.Value,
-        Texts.Exploring => ExploringCustom.Value,
-        Texts.Idle => IdleCustom.Value,
-        Texts.FightingInArena => FightingInArenaCustom.Value,
-        Texts.FightingBoss => FightingBossCustom.Value,
-        Texts.Singleplayer => SingleplayerCustom.Value,
-        Texts.Multiplayer => MultiplayerCustom.Value,
-        Texts.RaceUnknown => GetDefaultPresetText(text),
-        Texts.RaceImp => RaceImpCustom.Value,
-        Texts.RacePoon => RacePoonCustom.Value,
-        Texts.RaceKubold => RaceKuboldCustom.Value,
-        Texts.RaceByrdle => RaceByrdleCustom.Value,
-        Texts.RaceChang => RaceChangCustom.Value,
-        Texts.ClassUnknown => GetDefaultPresetText(text),
-        Texts.ClassNovice => ClassNoviceCustom.Value,
-        Texts.ClassFighter => ClassFighterCustom.Value,
-        Texts.ClassMystic => ClassMysticCustom.Value,
-        Texts.ClassBandit => ClassBanditCustom.Value,
-        _ => GetDefaultPresetText(text)
-    };
+        if (_customPresetData.TryGetValue(text, out var config))
+            return config.Value;
+
+        return GetDefaultPresetText(text);
+    }
 
     private string GetDefaultPresetText(Texts text) => text switch
     {
         Texts.PlayerAlive => $"Lv{{{LVL}}} {{{PLAYERRACEANDCLASS}}} ({{{HP}}}/{{{MAXHP}}} HP)",
         Texts.PlayerDead => $"Lv{{{LVL}}} {{{PLAYERRACEANDCLASS}}} (Fainted)",
         Texts.MainMenu => $"In Main Menu",
+        Texts.CharacterCreation => $"In Character Creation",
+        Texts.CharacterCreationDetails => $"Creating {{{CHARCREATERACE}}} named \"{{{CHARCREATENAME}}}\"",
         Texts.Exploring => $"{{{PLAYERNAME}}} exploring {{{WORLDAREA}}}",
         Texts.Idle => $"{{{PLAYERNAME}}} idle in {{{WORLDAREA}}}",
         Texts.FightingInArena => $"{{{PLAYERNAME}}} fighting in {{{WORLDAREA}}}",
-        Texts.FightingBoss => $"{{{PLAYERNAME}}} fighting a boss in {{{WORLDAREA}}}",
+        Texts.FightingBoss => $"{{{PLAYERNAME}}} fighting {{{BOSSNAME}}} in {{{WORLDAREA}}}",
+        Texts.DungeonBossEnd => $"{{{PLAYERNAME}}} getting boss loot in {{{WORLDAREA}}}",
         Texts.Singleplayer => $"Singleplayer",
         Texts.Multiplayer => $"Multiplayer on {{{SERVERNAME}}} ({{{PLAYERS}}}/{{{MAXPLAYERS}}})",
         Texts.RaceUnknown => "[Unknown Race]",
@@ -200,7 +212,37 @@ public class Display
         Texts.ClassFighter => "Fighter",
         Texts.ClassMystic => "Mystic",
         Texts.ClassBandit => "Bandit",
+        Texts.ClassTierBerserker => "Berserker",
+        Texts.ClassTierPaladin => "Paladin",
+        Texts.ClassTierMagus => "Magus",
+        Texts.ClassTierBishop => "Bishop",
+        Texts.ClassTierRogue => "Rogue",
+        Texts.ClassTierEngineer => "Engineer",
         _ => "[No Text]"
+    };
+
+    private string GetDetailedPresetText(Texts text) => text switch
+    {
+        Texts.PlayerAlive => $"Lv{{{LVL}}} {{{PLAYERRACEANDCLASS}}} ({{{EXP}}}/{{{EXPNEXT}}} EXP) ({{{HP}}}/{{{MAXHP}}} HP) ({{{MP}}}/{{{MAXMP}}} MP)",
+        Texts.PlayerDead => $"Lv{{{LVL}}} {{{PLAYERRACEANDCLASS}}} (Fainted) in {{{WORLDAREA}}}",
+        _ => GetDefaultPresetText(text)
+    };
+
+    private string GetDetailedPercentagesPresetText(Texts text) => text switch
+    {
+        Texts.PlayerAlive => $"Lv{{{LVL}}} {{{PLAYERRACEANDCLASS}}} ({{{EXPPCT}}}%) ({{{HPPCT}}}%) ({{{MPPCT}}}%)",
+        Texts.PlayerDead => $"Lv{{{LVL}}} {{{PLAYERRACEANDCLASS}}} (Fainted) in {{{WORLDAREA}}}",
+        _ => GetDefaultPresetText(text)
+    };
+
+    private string GetPonPresetText(Texts text) => text switch
+    {
+        Texts.RaceImp => "Inp",
+        Texts.RacePoon => "Pon",
+        Texts.RaceKubold => "Cubol",
+        Texts.RaceByrdle => "Birb",
+        Texts.RaceChang => "Tang",
+        _ => GetDefaultPresetText(text)
     };
 
     private string GetEmojisPresetText(Texts text) => text switch
@@ -208,23 +250,32 @@ public class Display
         Texts.PlayerAlive => $"{{{PLAYERRACE}}}{{{PLAYERCLASS}}} 🏆{{{LVL}}} ❤️{{{HP}}} ✨{{{MP}}}⚡{{{SP}}}",
         Texts.PlayerDead => $"{{{PLAYERRACE}}}{{{PLAYERCLASS}}} 🏆{{{LVL}}} 💀",
         Texts.MainMenu => $"📖 Main Menu",
+        Texts.CharacterCreation => $"📖 Character Creation",
+        Texts.CharacterCreationDetails => $"📖 Creating {{{CHARCREATERACE}}} \"{{{CHARCREATENAME}}}\"",
         Texts.Exploring => $"{{{PLAYERNAME}}} 🌎{{{WORLDAREA}}}",
         Texts.Idle => $"{{{PLAYERNAME}}} 🌎{{{WORLDAREA}}} 💤",
         Texts.FightingInArena => $"{{{PLAYERNAME}}} 🌎{{{WORLDAREA}}} ⚔️",
-        Texts.FightingBoss => $"{{{PLAYERNAME}}} 🌎{{{WORLDAREA}}} ⚔️👿",
+        Texts.FightingBoss => $"{{{PLAYERNAME}}} 🌎{{{WORLDAREA}}} ⚔️👿 {{{BOSSNAME}}}",
+        Texts.DungeonBossEnd => $"{{{PLAYERNAME}}} 🌎{{{WORLDAREA}}} ⚔️👿👑",
         Texts.Singleplayer => $"👤Solo",
         Texts.Multiplayer => $"👥{{{SERVERNAME}}} ({{{PLAYERS}}}/{{{MAXPLAYERS}}})",
         Texts.RaceUnknown => GetDefaultPresetText(text),
-        Texts.RaceImp => "👿",
-        Texts.RacePoon => "🐰",
-        Texts.RaceKubold => "🐲",
-        Texts.RaceByrdle => "🐦",
-        Texts.RaceChang => "🐿️",
+        Texts.RaceImp => "👿", // Angry face with horns
+        Texts.RacePoon => "🐰", // Rabbit
+        Texts.RaceKubold => "🐲", // Dragon face
+        Texts.RaceByrdle => "🐦", // Bird
+        Texts.RaceChang => "🐿️", // Chipmunk
         Texts.ClassUnknown => GetDefaultPresetText(text),
-        Texts.ClassNovice => "🌿",
+        Texts.ClassNovice => "🌿", // Herb
         Texts.ClassFighter => "🪓", // Axe
-        Texts.ClassMystic => "🌀",
-        Texts.ClassBandit => "👑",
+        Texts.ClassMystic => "🌀", // Cyclone
+        Texts.ClassBandit => "👑", // Crown
+        Texts.ClassTierBerserker => "🐗", // Boar
+        Texts.ClassTierPaladin => "🛡️", // Shield
+        Texts.ClassTierMagus => "🪄", // Magic wand
+        Texts.ClassTierBishop => "♗", // Bishop chess piece
+        Texts.ClassTierRogue => "🥷", // Ninja
+        Texts.ClassTierEngineer => "🔧", // Wrench
         _ => GetDefaultPresetText(text)
     };
 }